Properly print size_t for OutOfMemoryException
Summary:
%zu should be used to print size_t and when throwing
OutOfMemoryException, size_t is being cast to an int, losing the actual
value.

Reviewed By: mwilliams
Reviewers: mwilliams
CC: mwilliams
Differential Revision: 234805

+++ src/util/alloc.h
@@ -51,8 +51,8 @@ namespace HPHP {
 
 class OutOfMemoryException : public Exception {
 public:
-  OutOfMemoryException(int size)
-      : Exception("Unable to allocate %d bytes of memory", size) {}
+  OutOfMemoryException(size_t size)
+      : Exception("Unable to allocate %zu bytes of memory", size) {}
   virtual ~OutOfMemoryException() throw() {}
 };
